Game loop complet sauf score

+++ Assets/Scripts/Block/DiagBlock.cs
@@ -48,15 +48,15 @@ public class DiagBlock : PickUpDipositeBlock
         for (int i = 0; i < 3; i++)
         {
             image = feedback.transform.GetChild(0).GetChild(i).GetComponent<SpriteRenderer>();
-            if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == (int)RepairObject.CIRCUIT)
+            if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == RepairObject.CIRCUIT)
                 image.sprite = resourceSprite[0].Sprite;
-            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == (int)RepairObject.CLOU)
+            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == RepairObject.CLOU)
                 image.sprite = resourceSprite[1].Sprite;
-            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == (int)RepairObject.COLLE)
+            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == RepairObject.COLLE)
                 image.sprite = resourceSprite[2].Sprite;
-            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == (int)RepairObject.ENGRENAGE)
+            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == RepairObject.ENGRENAGE)
                 image.sprite = resourceSprite[3].Sprite;
-            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == (int)RepairObject.FIL)
+            else if (pickableObject.GetComponent<RepairableObject>().RecipeToDo[i] == RepairObject.FIL)
                 image.sprite = resourceSprite[4].Sprite;
         }
         Debug.Log(pickableObject.GetComponent<RepairableObject>().RecipeToDo[0]);
